Changed to use stol instead of stoi.

+++ Source/IntermediateCodeGenerator/TokenGenerator/LiteralTokenGenerator.cpp
@@ -6,7 +6,7 @@ ACC::LiteralTokenGenerator::LiteralTokenGenerator(ACC::ASTNode *node) : Expr(nod
 
 ACC::Dependency ACC::LiteralTokenGenerator::generate(ACC::Code &code) {
     Dependency out = code.createTemporary();
-    auto op = new Operator(OperatorId::ICOPY, static_cast<address>(std::stoi(node->str)), 0, out.temp);
+    auto op = new Operator(OperatorId::ICOPY, static_cast<address>(std::stol(node->str)), 0, out.temp);
     out.op = op;
     code.pushOp(op);
     return out;
